+++ index.js
@@ -51,16 +51,11 @@ class ScriptRunner {
   constructor(options) {
     this.options_ = options;
     this.aborted_ = false;
-    this.promises_ = [];
     this.logger_ = webdriver.logging.getLogger('wd-runjs.script-runner');
     this.logger_.debug(JSON.stringify(options));
   }
 
-  run(code) {
-    const STARTUP = [
-      'const ARGS = arguments;'
-    ];
-
+  async run(code, targets) {
     this.logger_.debug(`<<EOS\n${code}EOS`);
     const builder = new webdriver.Builder()
       .forBrowser(this.options_.browser);
@@ -68,87 +63,39 @@ class ScriptRunner {
       builder.usingServer(this.options_.server);
     }
     this.setBrowserOptions_(builder);
-    const promises = _.map(this.options_.uris, (uri) => {
+    return await Promise.all(targets.map(async (uri) => {
       let value = { uri: uri, browser: this.options_.browser };
       const driver = builder.build();
-      return driver
-        .then(() => {
-          if (this.options_.async) {
-            return driver.manage().setTimeouts({
-              script: this.options_.scriptTimeout * 1000
-            });
-          }
-          return Promise.resolve();
-        })
-        .then(this.abortable_(() => {
-          this.logger_.debug(`${uri}: start nativation...`);
-          return this.navigate_(driver, uri);
-        }))
-        .then(this.abortable_(() => {
-          this.logger_.debug(`${uri}: get the title...`);
-          return driver.getTitle();
-        }))
-        .then(this.abortable_((title) => {
-          value.title = title;
-          const script = _(STARTUP).concat(code).join('\n');
-          if (this.options_.async) {
-            this.logger_.debug(`${uri}: run the script asynchronously...`);
-            return driver.executeAsyncScript(
-              script, ...this.options_.scriptArgs);
-          } else {
-            this.logger_.debug(`${uri}: run the script synchronously...`);
-            return driver.executeScript(script, ...this.options_.scriptArgs);
-          }
-        }))
-        .then((result) => {
-          this.logger_.debug(`${uri}: done`);
-          value.result = result;
-        })
-        .catch((error) => {
-          this.logger_.debug(`${uri}: failed`);
-          value.error = error.message;
-        })
-        .then(() => {
-          this.logger_.debug(`${uri}: quit`);
-          return driver.quit();
-        })
-        .catch(() => {})  // ignore errors occurred in quit()
-        .then(() => value);
-    });
+      try {
+        await this.prepare_(driver);
+        await this.navigate_(driver, uri);
+        value.title = await this.getTitle_(driver, uri);
+        value.result = await this.executeScript_(driver, uri, code);
+        this.logger_.debug(`${uri}: done`);
+      } catch (error) {
+        this.logger_.debug(`${uri}: failed`);
+        value.error = error.message;
+      }
+      try {
+        this.logger_.debug(`${uri}: quit`);
+        await driver.quit();
+      } catch (e) {
+        // ignore errors occurred in quit()
+      }
 
-    this.promises_ = _.concat(this.promises_, promises);
-
-    return Promise.all(promises)
-      .then((results) => {  // always reach here even if errors occurred
-        // promises in the parent lexical scope are resolved.
-        // Remove them from `this.promises_`.
-        this.promises_ = _.difference(this.promises_, promises);
-        return results;
-      });
+      return value;
+    }));
   }
 
   abort() {
     this.logger_.debug('abort');
-    // The below code doesn't work as expected, espetially when the number of
-    // concurrency is larger than 1.
-    //
-    //   _.each(this.promises_, (promise) => promise.cancel('aborted'));
-    //
-    // Browser instances sometimes will not be closed even when quit() is called
-    // on all drivers.
-    //
-    // For avoiding the above situation, ScriptRunner checks `aborted_` flag
-    // before calling WebDriver's methods.
     this.aborted_ = true;
   }
 
-  abortable_(func) {
-    return _.rest((args) => {
-      if (this.aborted_) {
-        return Promise.reject(Error.wrap('aborted'));
-      }
-      return func.apply(this, args);
-    });
+  checkAborted_() {
+    if (this.aborted_) {
+      throw new Error('Aborted');
+    }
   }
 
   setBrowserOptions_(builder) {
@@ -161,21 +108,53 @@ class ScriptRunner {
     }
   }
 
-  navigate_(driver, uri) {
+  async prepare_(driver) {
+    if (this.options_.async) {
+      await driver.manage().setTimeouts({
+        script: this.options_.scriptTimeout * 1000
+      });
+    }
+  }
+
+  async navigate_(driver, uri) {
+    this.checkAborted_();
+    this.logger_.debug(`${uri}: start navigation...`);
     if (_.startsWith(uri, '@')) {
       const tab = _.trimStart(uri, '@');
       if (tab === 'current') {
         this.logger_.debug(`${uri}: use the current tab/window`);
-        return driver;  // noop
+        return;  // noop
       }
       throw new Error(`Unsupported window index: ${uri}`);
     }
     const parsedUri = url.parse(uri);
     if (_.isEmpty(parsedUri.protocol)) {
       const navigation = require(path.resolve(uri));
-      return navigation.navigate(driver);
+      await navigation.navigate(driver);
     } else {
-      return driver.get(uri);
+      await driver.get(uri);
+    }
+  }
+
+  async getTitle_(driver, uri) {
+    this.checkAborted_();
+    this.logger_.debug(`${uri}: get the title...`);
+    return await driver.getTitle();
+  }
+
+  async executeScript_(driver, uri, code) {
+    this.checkAborted_();
+    const script = [
+      'const ARGS = arguments;'
+    ].concat(code).join('\n');
+    if (this.options_.async) {
+      this.logger_.debug(`${uri}: run the script asynchronously...`);
+      return await driver.executeAsyncScript(
+        script, ...this.options_.scriptArgs);
+    } else {
+      this.logger_.debug(`${uri}: run the script synchronously...`);
+      return await driver.executeScript(
+        script, ...this.options_.scriptArgs);
     }
   }
 }